moved benchmark tests to `spec`
